Arreglado obtencion de datos closes #40 La obtencion de datos la hacia directamente desde la logica de negocio, se ha cambiado el lugar y se ha puesto en los tests para que asi asignacion de pedidos solo tenga codigo relacionado con la logica de negocio
